refactor: move app sources to src
