+++ public/index.html
@@ -6,7 +6,7 @@
     <title>Latent viewer</title>
 
     <!-- load our JS bundle -->
-    <script type="module" src="./main.ts"></script>
+    <script type="module" src="/src/bootstrap/index.ts"></script>
     <!-- <script type="module" src="./grid-integration.js"></script> -->
 
     <style>
