+++ src/components/LatentGrid.tsx
