+++ tests/modelSwitch.test.ts
@@ -19,7 +19,7 @@ vi.mock('react-dom/client', () => ({
   createRoot: () => ({ render: vi.fn() }),
 }));
 vi.mock(
-  '../public/LatentGrid',
+  '../src/components/LatentGrid',
   () => ({ LatentGrid: function LatentGrid() {}, LatentGridHandle: {} }),
   { virtual: true }
 );
@@ -133,7 +133,7 @@ beforeEach(async () => {
   pcAppEl = document.querySelector('pc-app') as HTMLElement & { app: any };
   (pcAppEl as any).app = app;
 
-  await import('../public/main.ts');
+  await import('../src/bootstrap/index.ts');
   document.dispatchEvent(new Event('DOMContentLoaded'));
   pcAppEl.dispatchEvent(new Event('ready'));
 });
